JWT verification for every route and post

+++ src/deploy.ts
@@ -1,4 +1,6 @@
 import { Hono } from "hono";
+import jwt from 'jsonwebtoken';
+import MongoService from "../structures/mongodb";
 import {exec, spawn} from "child_process";
 import { promisify } from 'util';
 import { streamText } from "hono/streaming";
@@ -9,7 +11,22 @@ const app = new Hono();
 app.post('/', async (c) => {
     const data = await c.req.text(); 
     const body = JSON.parse(data);
-    const { name, env, volumes, ports, network, repo_name, type } = body
+    const { name, env, volumes, ports, network, repo_name, type, token } = body
+    const client = MongoService.getInstance();
+    const jwtSecret = await client.findOne('vessyl', 'settings', {jwtSecret: {$exists : true}});
+    if (!jwtSecret) {
+        return c.text('JWT Secret not found');
+    }
+    let decoded : any = {};
+    try {
+        decoded = jwt.verify(token, jwtSecret.jwtSecret);
+    } catch (err) {
+        return c.text('Invalid token');
+    }
+    const user = await client.findOne('vessyl', 'users', {username: decoded.username});
+    if (!user) {
+        return c.text('User not found');
+    }
     let buildCommand = `docker run --rm -v /var/run/docker.sock:/var/run/docker.sock -v /var/run/docker.sock:/var/run/docker.sock -e TYPE=${type} -e REPO_NAME=${repo_name} vessyl-buildenv:0.0.18`
     return streamText(c, (stream) => {
         return new Promise((resolve, reject) => {
